+++ packages/react/.storybook/preview.tsx
@@ -7,10 +7,6 @@ import { themes, ensure } from "@storybook/theming";
 
 import "@fontsource-variable/open-sans/wdth.css";
 
-//import 'iframe-resizer';
-
-// require("iframe-resizer/js/iframeResizer.contentWindow");
-
 export const parameters = {
   controls: { expanded: true },
   docs: {