+++ packages/react/src/index.ts
@@ -109,7 +109,7 @@ export { default as NumberInput } from "./components/NumberInput";
 export { default as Pagination } from "./components/Pagination";
 export { default as RadioButton } from "./components/RadioButton";
 export { default as InputGroup } from "./components/InputGroup";
-export { default as RadioButtonGroup } from "./components/RadioButtonGroup";
+// export { default as RadioButtonGroup } from "./components/RadioButtonGroup";
 export { default as ReadMore } from "./components/ReadMore";
 //export { default as ReduxFormWrapper } from './components/ReduxFormWrapper';
 export { default as Search } from "./components/Search";