+++ packages/humanitarian-icons-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 11.14.50 (2023-10-09)
+
+**Note:** Version bump only for package @un/humanitarian-icons-react
+
+
+
+
+
+## 11.14.49 (2023-09-19)
+
+**Note:** Version bump only for package @un/humanitarian-icons-react
+
+
+
+
+
+## 11.14.48 (2023-09-19)
+
+**Note:** Version bump only for package @un/humanitarian-icons-react
+
+
+
+
+
 ## 11.14.47 (2023-08-07)
 
 **Note:** Version bump only for package @un/humanitarian-icons-react